+++ packages/samples/src/Camera.res.mjs
@@ -32,14 +32,12 @@ let cameraBounds = {
   }
 };
 
-let zeroVector = k.vec2(0, 0);
-
 let cameraVelocity = {
-  contents: zeroVector
+  contents: k.Vec2.ZERO
 };
 
 let lastTouchStart = {
-  contents: zeroVector
+  contents: k.Vec2.ZERO
 };
 
 let isDragging = {
@@ -104,17 +102,20 @@ let $$Map = {
 function onLoad() {
   let map = make$1();
   k.onTouchStart((pos, _touch) => {
-    lastTouchStart.contents = pos;
+    lastTouchStart.contents = k.toWorld(pos);
     isDragging.contents = true;
-    cameraVelocity.contents = zeroVector;
+    cameraVelocity.contents = k.Vec2.ZERO;
   });
   k.onTouchMove((pos, _touch) => {
     if (!isDragging.contents) {
       return;
     }
-    let delta = pos.sub(lastTouchStart.contents);
-    cameraVelocity.contents = delta.scale(k.vec2(-150, -100));
-    lastTouchStart.contents = pos;
+    let worldPos = k.toWorld(pos);
+    let delta = worldPos.sub(lastTouchStart.contents);
+    let sensitivity = k.vec2(-150, -100);
+    let worldDelta = delta.scale(sensitivity);
+    cameraVelocity.contents = worldDelta;
+    lastTouchStart.contents = worldPos;
   });
   k.onTouchEnd((param, _touch) => {
     isDragging.contents = false;
@@ -126,7 +127,7 @@ function onLoad() {
     let currentVelocity = cameraVelocity.contents;
     if (currentVelocity.len() > 0.1) {
       updateCamera(currentVelocity);
-      cameraVelocity.contents = currentVelocity.scale(k.vec2(0.9, 0.9));
+      cameraVelocity.contents = currentVelocity.scale(0.9);
       return;
     }
   });
@@ -148,10 +149,10 @@ function onLoad() {
           move = k.vec2(0, 10);
           break;
         default:
-          move = k.vec2(0, 0);
+          move = k.Vec2.ZERO;
       }
     } else {
-      move = k.vec2(0, 0);
+      move = k.Vec2.ZERO;
     }
     let result = currentCamPos.add(move);
     result.x = k.clamp(result.x, cameraBounds.x.min, cameraBounds.x.max);
@@ -179,7 +180,6 @@ export {
   gameWidth,
   gameHeight,
   cameraBounds,
-  zeroVector,
   cameraVelocity,
   lastTouchStart,
   isDragging,